feat: 유효성 검사 추가(없는 게시물을 조회/삭제 페이지가 없음을 알리는 메세지 반환 + 게시글 작성/수정할때 원래는 에러를 발생시켰는데 메세지 반환 방식으로 변경 + 게시글 검색할때 공백인 경우 예외처리) / refactor: 서비스가 아니라 컨트롤러에서 Entity->Dto로 바꾸도록 구조변경

+++ src/main/java/com/example/junhyuk_board/service/PostService.java
@@ -3,12 +3,14 @@ package com.example.junhyuk_board.service;
 import com.example.junhyuk_board.domain.Post;
 import com.example.junhyuk_board.domain.PostDTO;
 import com.example.junhyuk_board.repository.PostRepository;
+import com.example.junhyuk_board.Exception.ErrorCodes;
+import com.example.junhyuk_board.Exception.Customs.NoPostError;
+import com.example.junhyuk_board.Exception.Customs.UnableFormError;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -22,36 +24,38 @@ public class PostService {
         postRepository.save(post);
     }
 
-    public List<PostDTO> showPosts()
-    {
+    public List<PostDTO> showPosts() {
         List<Post> postList = postRepository.findTop100ByOrderByModifiedDate();
         List<PostDTO> postDTOList = postList.stream().map(post -> post.EntityToDTO()).collect(Collectors.toList());
         return postDTOList;
     }
 
-    public PostDTO findPostByID(Long id) {
-        Optional<Post> post = postRepository.findById(id);
-        Post naturalPost = post.get();
-        PostDTO postDTO = naturalPost.EntityToDTO();
-        return postDTO;
+    public Post findPostByID(Long id) {
+        if (!postRepository.existsById(id)) {
+            throw new NoPostError(ErrorCodes.NOT_FOUND);
+        }
+        return postRepository.findById(id).get();
     }
 
-    public PostDTO updatePostByID(Long id, String newTitle, String newContent) {
-        Optional<Post> post = postRepository.findById(id);
-        Post naturalPost = post.get();
-        naturalPost.setTitle(newTitle);
-        naturalPost.setContent(newContent);
-        postRepository.save(naturalPost);
-        PostDTO updatePost = naturalPost.EntityToDTO();
-        return updatePost;
+    public Post updatePostByID(Long id, String newTitle, String newContent) {
+        Post post = postRepository.findById(id).get();
+        post.setTitle(newTitle);
+        post.setContent(newContent);
+        postRepository.save(post);
+        return post;
     }
 
     public void deletePostByID(Long id) {
+        if (!postRepository.existsById(id)) {
+            throw new NoPostError(ErrorCodes.NOT_FOUND);
+        }
         postRepository.deleteById(id);
     }
 
-    public List<PostDTO> showPostBySearch(String Title)
-    {
+    public List<PostDTO> showPostBySearch(String Title) {
+        if (Title.length() < 1) {
+            throw new UnableFormError(ErrorCodes.BAD_REQUEST);
+        }
         List<Post> postList = postRepository.findTop100ByTitleContainingOrderByModifiedDate(Title);
         List<PostDTO> postDTOList = postList.stream().map(post -> post.EntityToDTO()).collect(Collectors.toList());
         return postDTOList;
